+++ aihub_web/aihub_web/nuxt.config.ts
@@ -85,6 +85,22 @@ export default defineNuxtConfig({
     },
   },
 
+  vite: {
+    optimizeDeps: {
+      // Disable discovery of new dependencies during dev to prevent page reloads
+      noDiscovery: true,
+      // Pre-bundle common dependencies
+      include: [
+        'vue',
+        'vue-router',
+        'pinia',
+        'primevue',
+        '@vueuse/core',
+        'vue-i18n',
+      ],
+    },
+  },
+
   eslint: {
     config: {
       stylistic: true,
@@ -138,22 +154,6 @@ export default defineNuxtConfig({
     },
     importTheme: {
       from: fileURLToPath(new URL('./themes/aihub-theme.ts', import.meta.url)),
-    },
-  },
-
-  vite: {
-    optimizeDeps: {
-      // Disable discovery of new dependencies during dev to prevent page reloads
-      noDiscovery: true,
-      // Pre-bundle common dependencies
-      include: [
-        'vue',
-        'vue-router',
-        'pinia',
-        'primevue',
-        '@vueuse/core',
-        'vue-i18n',
-      ],
     },
   },
