+++ aihub_web/aihub_web/pages/service/expert/groups/[id].vue
@@ -2,7 +2,7 @@
   <StructuralColumn
     :title="group?.name || t('expert.groups.group_detail')"
     :loading="isLoading"
-    :close-route="localePath('/service/expert?tab=groups')"
+    :close-route="localePath('/service/expert/groups')"
   >
     <div
       v-if="group"
